Fix sorting of nav section items
