+++ frontend/public/components/nav/perspective-nav.tsx
@@ -9,10 +9,10 @@ import {
   isNavSection,
   isNavItem,
   isSeparatorNavItem,
-  LoadedExtension,
 } from '@console/plugin-sdk';
 import { useActivePerspective, usePinnedResources } from '@console/shared';
 import { modelFor, referenceForModel } from '../../module/k8s';
+import { getSortedNavItems } from './navSortUtils';
 import confirmNavUnpinModal from './confirmNavUnpinModal';
 import { NavSection } from './section';
 import AdminNav from './admin-nav';
@@ -29,93 +29,6 @@ import './_perspective-nav.scss';
 const getLabelForResource = (resource: string): string => {
   const model = modelFor(resource);
   return model ? model.labelPlural : '';
-};
-
-const itemDependsOnItem = (
-  s1: LoadedExtension<PluginNavSection | NavItem | SeparatorNavItem>,
-  s2: LoadedExtension<PluginNavSection | NavItem | SeparatorNavItem>,
-) => {
-  if (!s1.properties.insertBefore && !s1.properties.insertAfter) {
-    return false;
-  }
-  const before = Array.isArray(s1.properties.insertBefore)
-    ? s1.properties.insertBefore
-    : [s1.properties.insertBefore];
-  const after = Array.isArray(s1.properties.insertAfter)
-    ? s1.properties.insertAfter
-    : [s1.properties.insertAfter];
-  return before.includes(s2.properties.id) || after.includes(s2.properties.id);
-};
-
-const isPositioned = (
-  item: LoadedExtension<PluginNavSection | NavItem | SeparatorNavItem>,
-  allItems: LoadedExtension<PluginNavSection | NavItem | SeparatorNavItem>[],
-) => !!allItems.find((i) => itemDependsOnItem(item, i));
-
-const findIndexForItem = (
-  item: LoadedExtension<PluginNavSection | NavItem | SeparatorNavItem>,
-  currentItems: LoadedExtension<PluginNavSection | NavItem | SeparatorNavItem>[],
-) => {
-  const { insertBefore, insertAfter } = item.properties;
-  let index = -1;
-  const before = Array.isArray(insertBefore) ? insertBefore : [insertBefore];
-  const after = Array.isArray(insertAfter) ? insertAfter : [insertAfter];
-  let count = 0;
-  while (count < before.length && index < 0) {
-    index = currentItems.findIndex((i) => i.properties.id === before[count]);
-    count++;
-  }
-  count = 0;
-  while (count < after.length && index < 0) {
-    index = currentItems.findIndex((i) => i.properties.id === after[count]);
-    if (index >= 0) {
-      index += 1;
-    }
-    count++;
-  }
-  return index;
-};
-
-const insertItem = (
-  item: LoadedExtension<PluginNavSection | NavItem | SeparatorNavItem>,
-  currentItems: LoadedExtension<PluginNavSection | NavItem | SeparatorNavItem>[],
-) => {
-  const index = findIndexForItem(item, currentItems);
-  if (index >= 0) {
-    currentItems.splice(index, 0, item);
-  } else {
-    currentItems.push(item);
-  }
-};
-
-const insertPositionedItems = (
-  insertItems: LoadedExtension<PluginNavSection | NavItem | SeparatorNavItem>[],
-  currentItems: LoadedExtension<PluginNavSection | NavItem | SeparatorNavItem>[],
-) => {
-  if (insertItems.length === 0) {
-    return;
-  }
-
-  const sortedItems = insertItems.filter((item) => !isPositioned(item, insertItems));
-  const positionedItems = insertItems.filter((item) => isPositioned(item, insertItems));
-
-  if (sortedItems.length === 0) {
-    // Circular dependencies
-    positionedItems.forEach((i) => insertItem(i, currentItems));
-    return;
-  }
-
-  sortedItems.forEach((i) => insertItem(i, currentItems));
-  insertPositionedItems(positionedItems, currentItems);
-};
-
-export const getSortedNavItems = (
-  topLevelItems: LoadedExtension<PluginNavSection | NavItem | SeparatorNavItem>[],
-) => {
-  const sortedItems = topLevelItems.filter((item) => !isPositioned(item, topLevelItems));
-  const positionedItems = topLevelItems.filter((item) => isPositioned(item, topLevelItems));
-  insertPositionedItems(positionedItems, sortedItems);
-  return sortedItems;
 };
 
 const PerspectiveNav: React.FC<{}> = () => {
